refactor: extract bp_fleet + bp_history to routes/fleet_history.py
Phase 5.10 of the incremental dashboard.py modularisation. Moves the 5
fleet routes and 7 history routes into a single new module — they're
conceptually related (ops/observability data layer) so they share a file,
same pattern as routes/alerts.py (bp_budget + bp_alerts).

Module-level helpers (_fleet_db, _fleet_db_lock, _fleet_check_key,
_fleet_update_statuses, _history_db, FLEET_HTML, _ext_emit,
AgentReliabilityScorer) stay in dashboard.py and are reached via late
`import dashboard as _d`. Pure mechanical move — zero behaviour change.

dashboard.py: 27,070 -> 26,815 (-255).
