+++ dashboard.py
@@ -74,6 +74,7 @@ from routes.alerts import bp_alerts, bp_budget
 from routes.channels import bp_channels
 from routes.overview import bp_overview
 from routes.components import bp_components
+from routes.fleet_history import bp_fleet, bp_history
 
 # History / time-series module
 try:
@@ -19232,10 +19233,10 @@ bp_auth = _Blueprint('auth', __name__)
 # bp_components moved to routes/components.py
 bp_config = _Blueprint('config', __name__)
 # bp_crons moved to routes/crons.py
-bp_fleet = _Blueprint('fleet', __name__)
+# bp_fleet moved to routes/fleet_history.py
 bp_gateway = _Blueprint('gateway', __name__)
 # bp_health moved to routes/health.py
-bp_history = _Blueprint('history', __name__)
+# bp_history moved to routes/fleet_history.py
 bp_logs = _Blueprint('logs', __name__)
 bp_memory = _Blueprint('memory', __name__)
 bp_otel = _Blueprint('otel', __name__)
@@ -20638,263 +20639,7 @@ load(); setInterval(load, 30000);
 """
 
 
-@bp_fleet.route("/fleet")
-def fleet_page():
-    """Fleet overview page for multi-node monitoring."""
-    return FLEET_HTML
-
-
-@bp_fleet.route("/api/nodes/register", methods=["POST"])
-def api_nodes_register():
-    """Register or update a remote node."""
-    if not _fleet_check_key(request):
-        return jsonify({"error": "Invalid or missing X-Fleet-Key"}), 401
-
-    data = request.get_json(silent=True) or {}
-    node_id = data.get("node_id", "").strip()
-    if not node_id:
-        return jsonify({"error": "node_id is required"}), 400
-
-    name = data.get("name", node_id)
-    hostname = data.get("hostname", "")
-    tags = json.dumps(data.get("tags", []))
-    version = data.get("version", "")
-    now = time.time()
-
-    with _fleet_db_lock:
-        db = _fleet_db()
-        db.execute(
-            """
-            INSERT INTO nodes (node_id, name, hostname, tags, version, registered_at, last_seen_at, status)
-            VALUES (?, ?, ?, ?, ?, ?, ?, 'online')
-            ON CONFLICT(node_id) DO UPDATE SET
-                name=excluded.name, hostname=excluded.hostname, tags=excluded.tags,
-                version=excluded.version, last_seen_at=excluded.last_seen_at, status='online'
-        """,
-            (node_id, name, hostname, tags, version, now, now),
-        )
-        db.commit()
-        db.close()
-
-    try:
-        _ext_emit("fleet.node_register", {"node_id": node_id})
-    except Exception:
-        pass
-    return jsonify({"ok": True, "node_id": node_id})
-
-
-@bp_fleet.route("/api/nodes/<node_id>/metrics", methods=["POST"])
-def api_nodes_push_metrics(node_id):
-    """Receive metrics push from a remote node."""
-    if not _fleet_check_key(request):
-        return jsonify({"error": "Invalid or missing X-Fleet-Key"}), 401
-
-    data = request.get_json(silent=True) or {}
-    now = time.time()
-
-    with _fleet_db_lock:
-        db = _fleet_db()
-        # Update last_seen
-        db.execute(
-            "UPDATE nodes SET last_seen_at = ?, status = 'online' WHERE node_id = ?",
-            (now, node_id),
-        )
-        # Store metrics snapshot
-        db.execute(
-            "INSERT INTO node_metrics (node_id, timestamp, metrics_json) VALUES (?, ?, ?)",
-            (node_id, now, json.dumps(data)),
-        )
-        db.commit()
-        db.close()
-
-    return jsonify({"ok": True, "received_at": now})
-
-
-@bp_fleet.route("/api/nodes")
-def api_nodes_list():
-    """List all registered nodes with latest metrics."""
-    _fleet_update_statuses()
-
-    with _fleet_db_lock:
-        db = _fleet_db()
-        nodes = db.execute("SELECT * FROM nodes ORDER BY name").fetchall()
-        result = []
-        total_cost = 0
-        total_sessions = 0
-        online_count = 0
-        offline_count = 0
-
-        for node in nodes:
-            n = dict(node)
-            n["tags"] = json.loads(n.get("tags") or "[]")
-
-            # Get latest metrics
-            row = db.execute(
-                "SELECT metrics_json FROM node_metrics WHERE node_id = ? ORDER BY timestamp DESC LIMIT 1",
-                (n["node_id"],),
-            ).fetchone()
-            n["latest_metrics"] = json.loads(row["metrics_json"]) if row else {}
-
-            # Aggregate stats
-            m = n["latest_metrics"]
-            if m.get("cost", {}).get("today_usd"):
-                total_cost += m["cost"]["today_usd"]
-            if m.get("sessions", {}).get("total_today"):
-                total_sessions += m["sessions"]["total_today"]
-
-            if n["status"] == "online":
-                online_count += 1
-            else:
-                offline_count += 1
-
-            # Remove internal fields
-            n.pop("api_key_hash", None)
-            result.append(n)
-
-        db.close()
-
-    return jsonify(
-        {
-            "nodes": result,
-            "fleet_summary": {
-                "total_nodes": len(result),
-                "online": online_count,
-                "offline": offline_count,
-                "total_cost_today": round(total_cost, 2),
-                "total_sessions_today": total_sessions,
-            },
-        }
-    )
-
-
-@bp_fleet.route("/api/nodes/<node_id>")
-def api_node_detail(node_id):
-    """Get detailed info for a single node with metric history."""
-    with _fleet_db_lock:
-        db = _fleet_db()
-        node = db.execute(
-            "SELECT * FROM nodes WHERE node_id = ?", (node_id,)
-        ).fetchone()
-        if not node:
-            db.close()
-            return jsonify({"error": "Node not found"}), 404
-
-        n = dict(node)
-        n["tags"] = json.loads(n.get("tags") or "[]")
-        n.pop("api_key_hash", None)
-
-        # Latest metrics
-        latest_row = db.execute(
-            "SELECT metrics_json FROM node_metrics WHERE node_id = ? ORDER BY timestamp DESC LIMIT 1",
-            (node_id,),
-        ).fetchone()
-        latest = json.loads(latest_row["metrics_json"]) if latest_row else {}
-
-        # 24h history
-        cutoff = time.time() - 86400
-        history_rows = db.execute(
-            "SELECT timestamp, metrics_json FROM node_metrics WHERE node_id = ? AND timestamp > ? ORDER BY timestamp",
-            (node_id, cutoff),
-        ).fetchall()
-        history = [
-            {"timestamp": r["timestamp"], "metrics": json.loads(r["metrics_json"])}
-            for r in history_rows
-        ]
-
-        db.close()
-
-    return jsonify(
-        {
-            "node": n,
-            "latest_metrics": latest,
-            "history": history,
-        }
-    )
-
-
-# ── Budget & Alert API Routes moved to routes/alerts.py ─────────────────
-
-
-# ── History / Time-Series API ────────────────────────────────────────────
-
-
-@bp_history.route("/api/history/metrics")
-def api_history_metrics():
-    """Query historical metrics. Params: metric, from, to, interval."""
-    if not _history_db:
-        return jsonify({"error": "History not available", "data": []}), 200
-    metric = request.args.get("metric", "tokens_in_total")
-    from_ts = request.args.get("from", type=float, default=time.time() - 3600)
-    to_ts = request.args.get("to", type=float, default=time.time())
-    interval = request.args.get("interval", None)
-    data = _history_db.query_metrics(metric, from_ts, to_ts, interval)
-    return jsonify({"data": data, "metric": metric})
-
-
-@bp_history.route("/api/history/metrics/list")
-def api_history_metrics_list():
-    """List available metric names."""
-    if not _history_db:
-        return jsonify({"metrics": []})
-    return jsonify({"metrics": _history_db.get_available_metrics()})
-
-
-@bp_history.route("/api/history/sessions")
-def api_history_sessions():
-    """Query historical session data."""
-    if not _history_db:
-        return jsonify({"data": []})
-    from_ts = request.args.get("from", type=float, default=time.time() - 3600)
-    to_ts = request.args.get("to", type=float, default=time.time())
-    session_key = request.args.get("session", None)
-    data = _history_db.query_sessions(from_ts, to_ts, session_key)
-    return jsonify({"data": data})
-
-
-@bp_history.route("/api/history/crons")
-def api_history_crons():
-    """Query historical cron run data."""
-    if not _history_db:
-        return jsonify({"data": []})
-    from_ts = request.args.get("from", type=float, default=time.time() - 3600)
-    to_ts = request.args.get("to", type=float, default=time.time())
-    job_id = request.args.get("job_id", None)
-    data = _history_db.query_crons(from_ts, to_ts, job_id)
-    return jsonify({"data": data})
-
-
-@bp_history.route("/api/history/snapshot/<float:timestamp>")
-def api_history_snapshot(timestamp):
-    """Get the snapshot closest to a given timestamp."""
-    if not _history_db:
-        return jsonify({"error": "History not available"}), 200
-    snap = _history_db.query_snapshot(timestamp)
-    if snap:
-        return jsonify(snap)
-    return jsonify({"error": "No snapshot found"}), 404
-
-
-@bp_history.route("/api/history/stats")
-def api_history_stats():
-    """Get history database stats."""
-    if not _history_db:
-        return jsonify({"enabled": False})
-    stats = _history_db.get_stats()
-    stats["enabled"] = True
-    return jsonify(stats)
-
-
-@bp_history.route("/api/history/reliability")
-def api_history_reliability():
-    """Cross-session behavioral reliability trend."""
-    if not _history_db:
-        return jsonify({"error": "History DB not available"}), 503
-    from history import AgentReliabilityScorer
-
-    scorer = AgentReliabilityScorer(_history_db)
-    window = request.args.get("window", 30, type=int)
-    result = scorer.score(window_days=window)
-    return jsonify(result)
+# ── Fleet + History API Routes moved to routes/fleet_history.py ─────────
 
 
 # ── Billing Mode Heuristics (API key vs OAuth/included) ──────────────────
